ATS/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pdfplumber==0.11.4
google-generativeai==0.8.5
python-dotenv==1.1.1
protobuf>=5.28.0
